Measurement Unit changed to double values where possible

+++ BeehiveMeasurement/LightSensor.h
@@ -22,7 +22,7 @@ public:
 
 private:
   void requestData();
-  uint16_t readData();
+  double readData();
 
   IMeasurementSink& sink;
   int interval;
